clean up, add xyzrgb radial
